Added code to product-routes.js to update the product.

+++ routes/api/product-routes.js
@@ -51,48 +51,38 @@ router.post('/products', async (req, res) => {
 });
 
 // update product
-router.put('/:id', (req, res) => {
-  // update product data
-  Product.update(req.body, {
-    where: {
-		id: req.params.id,
-    },
-  })
-    .then((product) => {
+router.put('/products/:id', async (req, res) => {
+	try {
+		const [affectedRows] = await Product.update(req.body, {
+			where: { id: req.params.id },
+		});
+		if (affectedRows <= 0) {
+			res.status(404).json({ message: 'No product found with this id' });
+			return;
+		}
 		if (req.body.tagIds && req.body.tagIds.length) {
-
-			ProductTag.findAll({
-			where: { product_id: req.params.id }
-			}).then((productTags) => {
-			// create filtered list of new tag_ids
+			const productTags = await ProductTag.findAll({
+			where: { product_id: req.params.id },
+			});
 			const productTagIds = productTags.map(({ tag_id }) => tag_id);
 			const newProductTags = req.body.tagIds
-				.filter((tag_id) => !productTagIds.includes(tag_id))
-				.map((tag_id) => {
+			.filter((tag_id) => !productTagIds.includes(tag_id))
+			.map((tag_id) => {
 				return {
-					product_id: req.params.id,
-					tag_id,
+				product_id: req.params.id,
+				tag_id,
 				};
-				});
-
-			// figure out which ones to remove
-			const productTagsToRemove = productTags
-				.filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
-				.map(({ id }) => id);
-			// run both actions
-			return Promise.all([
-				ProductTag.destroy({ where: { id: productTagsToRemove } }),
-				ProductTag.bulkCreate(newProductTags),
-			]);
 			});
+			const productTagsToRemove = productTags
+			.filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
+			.map(({ id }) => id);
+			await ProductTag.destroy({ where: { id: productTagsToRemove } });
+			await ProductTag.bulkCreate(newProductTags);
 		}
-
-		return res.json(product);
-    })
-    .catch((err) => {
-		// console.log(err);
-		res.status(400).json(err);
-    });
+		res.status(200).json({ message: 'Product updated successfully' });
+	} catch (err) {
+	  	res.status(400).json(err);
+	}
 });
 
 router.delete('/products/:id', async (req, res) => {
